Yet another of my potentially useless projects.
So, here we have the implementation of a userspace
driver that uses libusb from computerquip/libusb on github, and
uinput from the linux input subsystem for userspace.

Things that are left to do are as follows:
1) Any output which includes:
	a) LED packets
	b) Rumble packets

These are easy to do, but synchronizing them may be difficult.

I do eventually plan to support more than one xbox controller.